+++ CHANGELOG.md
@@ -20,6 +20,7 @@
     - `credentials`
     - `encryption`
 - Added support for `Session.directory` and `Session.read.directory` to retrieve the list of all files on a stage with metadata.
+- Added support for `FileOperation.copy_files` to copy files from a source location to an output stage.
 
 - Added support for the following scalar functions in `functions.py`:
   - `all_user_names`